(add) merge positions / select position styles
