+++ src/pages/MergePositions/index.tsx
@@ -1,7 +1,7 @@
-import { Contents } from 'components/MergePositions/Contents'
 import { PageTitle } from 'components/pureStyledComponents/PageTitle'
 import { ConditionProvider } from 'contexts/ConditionContext'
 import { MultiPositionsProvider } from 'contexts/MultiPositionsContext'
+import { Contents } from 'pages/MergePositions/Contents'
 import React from 'react'
 
 export const MergePositions = () => {
@@ -9,7 +9,6 @@ export const MergePositions = () => {
     <MultiPositionsProvider checkForEmptyBalance={true}>
       <ConditionProvider checkForConditionNotResolved={true}>
         <PageTitle>Merge Positions</PageTitle>
-        {/* {loading && <InlineLoading />} */}
         <Contents />
       </ConditionProvider>
     </MultiPositionsProvider>
